MAINT: Move `exec_` to utils
